- Got code working under MSVC++. -- Added VS2015 project. -- Added redirection of cout/cerr/clog to Visual Studio 'Output' window -- Fixed bug with natural_sort -- Worked around bug with tbb::task_group_context::register_pending_exception - Removed tabs from source code. - Updated to Lua 5.3 - Added unit-test names - Removed unused C++ triggersystem code.

+++ trunk/src/AudioGlitz.h
@@ -19,7 +19,7 @@ public:
     {}
 
     virtual void display(LayeredCanvas &canvas) const override {
-        canvas.playSound(key, n);
+        canvas.playSound(key, static_cast<int>(n));
     }
     virtual std::size_t clone_size() const override {
         return sizeof *this;
